Improve the implementation
