+++ frontend/app/utils/usernameGenerator.ts
@@ -1,0 +1,958 @@
+// Local word lists for username generation
+export const adjectives = [
+  "Swift",
+  "Bright",
+  "Bold",
+  "Quick",
+  "Smart",
+  "Cool",
+  "Fast",
+  "Wild",
+  "Free",
+  "Pure",
+  "Strong",
+  "Brave",
+  "Calm",
+  "Sharp",
+  "Wise",
+  "Kind",
+  "True",
+  "Fair",
+  "Deep",
+  "Clear",
+  "Smooth",
+  "Solid",
+  "Noble",
+  "Proud",
+  "Lucky",
+  "Happy",
+  "Golden",
+  "Silver",
+  "Cosmic",
+  "Digital",
+  "Crypto",
+  "Quantum",
+  "Decentralized",
+  "Trustless",
+  "Immutable",
+  "Permissionless",
+  "Atomic",
+  "Zero",
+  "Alpha",
+  "Beta",
+  "Genesis",
+  "Final",
+  "Ultra",
+  "Meta",
+  "Super",
+  "Hyper",
+  "Mega",
+  "Giga",
+  "Nano",
+  "Micro",
+  "Stealth",
+  "Phantom",
+  "Shadow",
+  "Elite",
+  "Prime",
+  "Apex",
+  "Void",
+  "Dark",
+  "Light",
+  "Neon",
+  "Electric",
+  "Plasma",
+  "Laser",
+  "Turbo",
+  "Blaze",
+  "Frost",
+  "Storm",
+  "Cyber",
+  "Neural",
+  "Binary",
+  "Flux",
+  "Nexus",
+  "Vector",
+  "Matrix",
+  "Prism",
+  "Pulse",
+  "Echo",
+  "Viral",
+  "Ghost",
+  "Rogue",
+  "Mystic",
+  "Sacred",
+  "Divine",
+  "Eternal",
+  "Infinite",
+  "Stellar",
+  "Galactic",
+  "Lunar",
+  "Solar",
+  "Orbital",
+  "Radiant",
+  "Luminous",
+  "Brilliant",
+  "Shining",
+  "Glowing",
+  "Blazing",
+  "Burning",
+  "Frozen",
+  "Crystal",
+  "Diamond",
+  "Platinum",
+  "Titanium",
+  "Carbon",
+  "Iron",
+  "Steel",
+  "Copper",
+  "Bronze",
+  "Emerald",
+  "Ruby",
+  "Sapphire",
+  "Obsidian",
+  "Granite",
+  "Marble",
+  "Quartz",
+  "Jade",
+  "Amber",
+  "Pearl",
+  "Opal",
+  "Topaz",
+  "Crimson",
+  "Scarlet",
+  "Azure",
+  "Violet",
+  "Indigo",
+  "Magenta",
+  "Cyan",
+  "Teal",
+  "Lime",
+  "Orange",
+  "Purple",
+  "Maroon",
+  "Navy",
+  "Olive",
+  "Pink",
+  "Brown",
+  "Gray",
+  "Black",
+  "White",
+  "Red",
+  "Blue",
+  "Green",
+  "Yellow",
+  "Arctic",
+  "Polar",
+  "Tropical",
+  "Desert",
+  "Mountain",
+  "Forest",
+  "Ocean",
+  "Sky",
+  "Earth",
+  "Fire",
+  "Water",
+  "Air",
+  "Space",
+  "Time",
+  "Speed",
+  "Power",
+  "Energy",
+  "Force",
+  "Chaos",
+  "Order",
+  "Peace",
+  "War",
+  "Love",
+  "Hate",
+  "Hope",
+  "Fear",
+  "Joy",
+  "Rage",
+  "Calm",
+  "Wild",
+  "Tame",
+  "Lost",
+  "Found",
+  "Hidden",
+  "Visible",
+  "Secret",
+  "Open",
+  "Closed",
+  "Locked",
+  "Free",
+  "Bound",
+  "Broken",
+  "Fixed",
+  "New",
+  "Old",
+  "Young",
+  "Ancient",
+  "Modern",
+  "Future",
+  "Past",
+  "Present",
+  "Real",
+  "Virtual",
+  "Artificial",
+  "Natural",
+  "Synthetic",
+  "Organic",
+  "Toxic",
+  "Pure",
+  "Clean",
+  "Dirty",
+  "Fresh",
+  "Stale",
+  "Hot",
+  "Cold",
+  "Warm",
+  "Cool",
+];
+
+export const nouns = [
+  "Tiger",
+  "Eagle",
+  "Wolf",
+  "Lion",
+  "Bear",
+  "Fox",
+  "Hawk",
+  "Shark",
+  "Dragon",
+  "Phoenix",
+  "Storm",
+  "Thunder",
+  "Lightning",
+  "Fire",
+  "Ice",
+  "Star",
+  "Moon",
+  "Sun",
+  "Wind",
+  "Rain",
+  "Mountain",
+  "River",
+  "Ocean",
+  "Forest",
+  "Desert",
+  "Valley",
+  "Peak",
+  "Wave",
+  "Stone",
+  "Crystal",
+  "Block",
+  "Chain",
+  "Node",
+  "Hash",
+  "Token",
+  "Coin",
+  "Wallet",
+  "Key",
+  "Proof",
+  "Stake",
+  "Mint",
+  "Fork",
+  "Bridge",
+  "Pool",
+  "Swap",
+  "Yield",
+  "Vault",
+  "Protocol",
+  "Layer",
+  "Network",
+  "Gas",
+  "Fee",
+  "Burn",
+  "Lock",
+  "Shard",
+  "Merkle",
+  "Nonce",
+  "Oracle",
+  "Rollup",
+  "Zap",
+  "Panther",
+  "Falcon",
+  "Cobra",
+  "Viper",
+  "Raven",
+  "Serpent",
+  "Lynx",
+  "Jaguar",
+  "Leopard",
+  "Cheetah",
+  "Rhino",
+  "Buffalo",
+  "Bison",
+  "Mammoth",
+  "Kraken",
+  "Leviathan",
+  "Hydra",
+  "Griffin",
+  "Wyvern",
+  "Unicorn",
+  "Pegasus",
+  "Sphinx",
+  "Minotaur",
+  "Titan",
+  "Golem",
+  "Wraith",
+  "Specter",
+  "Phantom",
+  "Spirit",
+  "Soul",
+  "Essence",
+  "Core",
+  "Heart",
+  "Mind",
+  "Brain",
+  "Nexus",
+  "Portal",
+  "Gateway",
+  "Beacon",
+  "Signal",
+  "Pulse",
+  "Echo",
+  "Wave",
+  "Flow",
+  "Stream",
+  "Current",
+  "Surge",
+  "Spark",
+  "Flash",
+  "Bolt",
+  "Ray",
+  "Beam",
+  "Laser",
+  "Plasma",
+  "Energy",
+  "Power",
+  "Force",
+  "Field",
+  "Zone",
+  "Realm",
+  "Domain",
+  "Empire",
+  "Kingdom",
+  "Throne",
+  "Crown",
+  "Scepter",
+  "Blade",
+  "Sword",
+  "Shield",
+  "Armor",
+  "Forge",
+  "Hammer",
+  "Anvil",
+  "Axe",
+  "Spear",
+  "Arrow",
+  "Bow",
+  "Staff",
+  "Wand",
+  "Rod",
+  "Club",
+  "Mace",
+  "Dagger",
+  "Knife",
+  "Scythe",
+  "Whip",
+  "Chain",
+  "Rope",
+  "Net",
+  "Trap",
+  "Cage",
+  "Prison",
+  "Tower",
+  "Castle",
+  "Fortress",
+  "Citadel",
+  "Bastion",
+  "Keep",
+  "Wall",
+  "Gate",
+  "Door",
+  "Window",
+  "Bridge",
+  "Path",
+  "Road",
+  "Trail",
+  "Journey",
+  "Quest",
+  "Mission",
+  "Task",
+  "Goal",
+  "Dream",
+  "Vision",
+  "Hope",
+  "Wish",
+  "Desire",
+  "Want",
+  "Need",
+  "Fear",
+  "Doubt",
+  "Faith",
+  "Trust",
+  "Belief",
+  "Truth",
+  "Lie",
+  "Secret",
+  "Mystery",
+  "Riddle",
+  "Puzzle",
+  "Code",
+  "Cipher",
+  "Symbol",
+  "Sign",
+  "Mark",
+  "Stamp",
+  "Seal",
+  "Badge",
+  "Medal",
+  "Trophy",
+  "Prize",
+  "Reward",
+  "Gift",
+  "Present",
+  "Surprise",
+  "Wonder",
+  "Miracle",
+  "Magic",
+  "Spell",
+  "Charm",
+  "Curse",
+  "Blessing",
+  "Prayer",
+  "Ritual",
+  "Ceremony",
+  "Festival",
+  "Party",
+  "Dance",
+  "Song",
+  "Music",
+  "Sound",
+  "Voice",
+  "Word",
+  "Name",
+  "Title",
+  "Rank",
+  "Level",
+  "Grade",
+  "Class",
+  "Type",
+  "Kind",
+  "Sort",
+  "Form",
+  "Shape",
+  "Size",
+  "Color",
+  "Shade",
+  "Tone",
+  "Hue",
+  "Tint",
+  "Light",
+  "Dark",
+  "Shadow",
+  "Glow",
+  "Shine",
+  "Spark",
+  "Flame",
+  "Ember",
+  "Ash",
+  "Dust",
+  "Sand",
+  "Dirt",
+  "Mud",
+  "Clay",
+  "Rock",
+  "Boulder",
+  "Pebble",
+  "Gem",
+  "Jewel",
+  "Diamond",
+  "Ruby",
+  "Emerald",
+  "Sapphire",
+  "Topaz",
+  "Opal",
+  "Pearl",
+  "Coral",
+  "Shell",
+  "Bone",
+  "Skull",
+  "Spine",
+  "Rib",
+  "Tooth",
+  "Fang",
+  "Claw",
+  "Talon",
+  "Wing",
+  "Feather",
+  "Scale",
+  "Fur",
+  "Hair",
+  "Skin",
+  "Hide",
+  "Leather",
+  "Fabric",
+  "Cloth",
+  "Silk",
+  "Cotton",
+  "Wool",
+  "Linen",
+  "Velvet",
+  "Satin",
+  "Lace",
+  "Thread",
+  "String",
+  "Wire",
+  "Cable",
+  "Cord",
+  "Line",
+  "Loop",
+  "Circle",
+  "Ring",
+  "Band",
+  "Strip",
+  "Belt",
+  "Strap",
+  "Buckle",
+  "Button",
+  "Zipper",
+  "Hook",
+  "Latch",
+  "Lock",
+  "Key",
+  "Handle",
+  "Grip",
+  "Hold",
+  "Clutch",
+  "Grasp",
+  "Touch",
+  "Feel",
+  "Sense",
+  "Sight",
+  "Sound",
+  "Taste",
+  "Smell",
+  "Memory",
+  "Thought",
+  "Idea",
+  "Concept",
+  "Theory",
+  "Fact",
+  "Data",
+  "Info",
+  "Knowledge",
+  "Wisdom",
+  "Intelligence",
+  "Genius",
+  "Talent",
+  "Skill",
+  "Art",
+  "Craft",
+  "Trade",
+  "Work",
+  "Job",
+  "Task",
+  "Duty",
+  "Role",
+  "Part",
+  "Piece",
+  "Bit",
+  "Chunk",
+  "Block",
+  "Unit",
+  "Element",
+  "Component",
+  "Part",
+  "Section",
+  "Division",
+  "Branch",
+  "Arm",
+  "Leg",
+  "Foot",
+  "Hand",
+  "Finger",
+  "Thumb",
+  "Palm",
+  "Wrist",
+  "Elbow",
+  "Shoulder",
+  "Neck",
+  "Head",
+  "Face",
+  "Eye",
+  "Ear",
+  "Nose",
+  "Mouth",
+  "Lip",
+  "Tongue",
+  "Teeth",
+  "Chin",
+  "Cheek",
+  "Brow",
+  "Forehead",
+  "Temple",
+  "Jaw",
+  "Beard",
+  "Mustache",
+  "Smile",
+  "Frown",
+  "Grin",
+  "Laugh",
+  "Cry",
+  "Tear",
+  "Drop",
+  "Splash",
+  "Ripple",
+  "Tide",
+  "Current",
+  "Flow",
+  "Stream",
+  "River",
+  "Lake",
+  "Pond",
+  "Pool",
+  "Sea",
+  "Ocean",
+  "Bay",
+  "Gulf",
+  "Strait",
+  "Channel",
+  "Harbor",
+  "Port",
+  "Dock",
+  "Pier",
+  "Wharf",
+  "Shore",
+  "Beach",
+  "Coast",
+  "Island",
+  "Peninsula",
+  "Cape",
+  "Point",
+  "Cliff",
+  "Rock",
+  "Cave",
+  "Cavern",
+  "Tunnel",
+  "Mine",
+  "Shaft",
+  "Pit",
+  "Hole",
+  "Gap",
+  "Space",
+  "Room",
+  "Hall",
+  "Chamber",
+  "Vault",
+  "Cellar",
+  "Basement",
+  "Attic",
+  "Loft",
+  "Floor",
+  "Ceiling",
+  "Wall",
+  "Corner",
+  "Edge",
+  "Border",
+  "Boundary",
+  "Limit",
+  "End",
+  "Start",
+  "Beginning",
+  "Origin",
+  "Source",
+  "Root",
+  "Base",
+  "Foundation",
+  "Ground",
+  "Earth",
+  "Soil",
+  "Land",
+  "Field",
+  "Meadow",
+  "Prairie",
+  "Plain",
+  "Plateau",
+  "Hill",
+  "Mountain",
+  "Peak",
+  "Summit",
+  "Top",
+  "Crown",
+  "Apex",
+  "Zenith",
+  "Height",
+  "Altitude",
+  "Depth",
+  "Bottom",
+  "Floor",
+  "Basement",
+  "Foundation",
+  "Core",
+  "Center",
+  "Middle",
+  "Hub",
+  "Axis",
+  "Pivot",
+  "Hinge",
+  "Joint",
+  "Connection",
+  "Link",
+  "Bond",
+  "Tie",
+  "Knot",
+  "Twist",
+  "Turn",
+  "Bend",
+  "Curve",
+  "Arc",
+  "Circle",
+  "Spiral",
+  "Helix",
+  "Coil",
+  "Spring",
+  "Bounce",
+  "Jump",
+  "Leap",
+  "Hop",
+  "Skip",
+  "Step",
+  "Stride",
+  "Walk",
+  "Run",
+  "Sprint",
+  "Dash",
+  "Rush",
+  "Speed",
+  "Velocity",
+  "Momentum",
+  "Inertia",
+  "Motion",
+  "Movement",
+  "Action",
+  "Activity",
+  "Operation",
+  "Function",
+  "Process",
+  "Method",
+  "Way",
+  "Path",
+  "Route",
+  "Course",
+  "Direction",
+  "Heading",
+  "Bearing",
+  "Angle",
+  "Degree",
+  "Measure",
+  "Scale",
+  "Size",
+  "Length",
+  "Width",
+  "Height",
+  "Depth",
+  "Volume",
+  "Mass",
+  "Weight",
+  "Density",
+  "Pressure",
+  "Force",
+  "Power",
+  "Energy",
+  "Strength",
+  "Might",
+  "Vigor",
+  "Vitality",
+  "Life",
+  "Living",
+  "Being",
+  "Existence",
+  "Reality",
+  "Truth",
+  "Fact",
+  "Matter",
+  "Substance",
+  "Material",
+  "Object",
+  "Thing",
+  "Item",
+  "Article",
+  "Piece",
+  "Part",
+  "Fragment",
+  "Shard",
+  "Splinter",
+  "Chip",
+  "Crack",
+  "Break",
+  "Split",
+  "Tear",
+  "Rip",
+  "Cut",
+  "Slice",
+  "Chop",
+  "Hack",
+  "Strike",
+  "Hit",
+  "Blow",
+  "Impact",
+  "Crash",
+  "Smash",
+  "Crush",
+  "Squeeze",
+  "Press",
+  "Push",
+  "Pull",
+  "Drag",
+  "Lift",
+  "Raise",
+  "Lower",
+  "Drop",
+  "Fall",
+  "Descent",
+  "Ascent",
+  "Rise",
+  "Climb",
+  "Reach",
+  "Stretch",
+  "Extend",
+  "Expand",
+  "Grow",
+  "Increase",
+  "Decrease",
+  "Reduce",
+  "Shrink",
+  "Contract",
+  "Compress",
+  "Compact",
+  "Dense",
+  "Thick",
+  "Thin",
+  "Narrow",
+  "Wide",
+  "Broad",
+  "Long",
+  "Short",
+  "Tall",
+  "Small",
+  "Large",
+  "Big",
+  "Huge",
+  "Giant",
+  "Massive",
+  "Enormous",
+  "Tiny",
+  "Minute",
+  "Micro",
+  "Mini",
+  "Mega",
+  "Ultra",
+  "Super",
+  "Hyper",
+  "Extra",
+  "Special",
+  "Unique",
+  "Rare",
+  "Common",
+  "Normal",
+  "Regular",
+  "Standard",
+  "Basic",
+  "Simple",
+  "Complex",
+  "Complicated",
+  "Difficult",
+  "Easy",
+  "Hard",
+  "Soft",
+  "Smooth",
+  "Rough",
+  "Sharp",
+  "Dull",
+  "Bright",
+  "Dark",
+  "Light",
+  "Heavy",
+  "Fast",
+  "Slow",
+  "Quick",
+  "Swift",
+  "Rapid",
+  "Instant",
+  "Immediate",
+  "Sudden",
+  "Gradual",
+  "Steady",
+  "Stable",
+  "Solid",
+  "Firm",
+  "Fixed",
+  "Loose",
+  "Tight",
+  "Open",
+  "Closed",
+  "Full",
+  "Empty",
+  "Complete",
+  "Partial",
+  "Whole",
+  "Half",
+  "Quarter",
+  "Third",
+  "Double",
+  "Triple",
+  "Single",
+  "Multiple",
+  "Many",
+  "Few",
+  "Several",
+  "Some",
+  "All",
+  "None",
+  "Zero",
+  "One",
+  "Two",
+  "Three",
+  "Four",
+  "Five",
+  "Six",
+  "Seven",
+  "Eight",
+  "Nine",
+  "Ten",
+  "Hundred",
+  "Thousand",
+  "Million",
+  "Billion",
+  "Infinity",
+];
+
+export function generateRandomUsername(): string {
+  const randomAdjective =
+    adjectives[Math.floor(Math.random() * adjectives.length)];
+  const randomNoun = nouns[Math.floor(Math.random() * nouns.length)];
+
+  let combinedUsername = randomAdjective + randomNoun;
+
+  // Ensure max 31 characters
+  if (combinedUsername.length > 31) {
+    // Try to trim the noun first
+    const maxNounLength = 31 - randomAdjective.length;
+    if (maxNounLength > 3) {
+      combinedUsername =
+        randomAdjective + randomNoun.substring(0, maxNounLength);
+    } else {
+      // If still too long, trim both words proportionally
+      const halfLength = Math.floor(31 / 2);
+      combinedUsername =
+        randomAdjective.substring(0, halfLength) +
+        randomNoun.substring(0, 31 - halfLength);
+    }
+  }
+
+  return combinedUsername;
+}
